+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 * added a way to register an object to be cleaned up via MemoryPressureManager
 * fixed get_own_cpu_usage() to work on Windows
 * added __len__ to Optional
+* major bugfix in Optional